requirements.txt
# Web App
fastapi > 0.100.0
pydantic > 2.0.0
pydantic-settings
uvicorn[standard]
requests >= 2.31.0


# Environment Variables + Settings
python-dotenv
toml>=0.10.2


mongomock
# Testing
pytest
pytest-asyncio
httpx

# GCP
google-cloud-core==2.2.2
google-cloud-logging==3.0.0
regex

# Linting
